modified:   package-lock.json 	modified:   src/app/(profile-related)/profile/actions.ts 	deleted:    src/app/(profile-related)/profile/view-photo/page.tsx 	new file:   src/app/(profile-related)/view-photo/page.tsx
